Final implementation of MixMatch algorithm
