+++ semisupervised/__init__.py
@@ -1,0 +1,1 @@
+""" Classes and functions use across different Semi-Supervised algorithms """
