+++ semisupervised/mixmatch/__init__.py
@@ -1,0 +1,12 @@
+""" 
+Some ideas taken from [YU1ut repository](https://github.com/YU1ut/MixMatch-pytorch) and others from [noachr repository](https://github.com/noachr/MixMatch-fastai)
+## References
+```
+@article{berthelot2019mixmatch,
+  title={MixMatch: A Holistic Approach to Semi-Supervised Learning},
+  author={Berthelot, David and Carlini, Nicholas and Goodfellow, Ian and Papernot, Nicolas and Oliver, Avital and Raffel, Colin},
+  journal={arXiv preprint arXiv:1905.02249},
+  year={2019}
+}
+```
+"""
