+++ semisupervised/mixmatch/callback.py
@@ -3,9 +3,10 @@ import torch
 from fastai.callback.core import Callback
 from fastai.callback.mixup import MixUp
 
-from mixmatch.utils import interleave
+from semisupervised.utils import interleave
 
 class MixMatchCallback(Callback):
+    """ MixMatch required preprocess before each batch """
 
     run_after = MixUp
     def __init__(self, unlabel_dl, transform_dl, T):
@@ -14,14 +15,17 @@ class MixMatchCallback(Callback):
         self.T = T
 
     def before_batch(self):
+
+        # Only process if is trainig
         if not self.training: return
         
+        # Extract a batch of unlabel images and repeat twice the transformation (different results due to randomness)
         raw_inputs_u = self.unlabel_dl.one_batch()[0]
         inputs_u = self.transform_dl.after_batch(raw_inputs_u.detach())
         inputs_u2 = self.transform_dl.after_batch(raw_inputs_u.detach())
 
+        # Compute guessed labels of unlabel samples
         with torch.no_grad():
-            # compute guessed labels of unlabel samples
             outputs_u = self.model(inputs_u)
             outputs_u2 = self.model(inputs_u2)
             p = (torch.softmax(outputs_u, dim=1) + torch.softmax(outputs_u2, dim=1)) / 2
@@ -29,16 +33,19 @@ class MixMatchCallback(Callback):
             targets_u = pt / pt.sum(dim=1, keepdim=True)
             targets_u = targets_u.detach()
         
+        # Make suree the input has the right structure
         input_x = self.learn.xb
         if type(input_x) is tuple:
             input_x = input_x[0]
 
+        # Make suree the target has the right structure
         target_x = self.learn.yb
         if type(target_x) is tuple:
             target_x = target_x[0]
     
+        # Set together label and unlabel data
         self.learn.xb = torch.cat([input_x, inputs_u, inputs_u2], dim=0).unsqueeze(0)
         self.learn.yb = torch.cat([target_x, targets_u, targets_u], dim=0).unsqueeze(0)
 
-        # interleave labeled and unlabed samples between batches to get correct batchnorm calculation 
+        # Interleave labeled and unlabed samples between batches to get correct batchnorm calculation 
         self.learn.xb = interleave(self.learn.xb)
